Updated physics logic

+++ a3/Edge.pde
@@ -2,13 +2,10 @@ class Edge {
   Node node1, node2;
   float curr_length, resting_length;
   
-  Edge(Node _node1, Node _node2, int _resting_length){
+  Edge(Node _node1, Node _node2, float _resting_length){
     node1 = _node1;
     node2 = _node2;
     resting_length = _resting_length;
   }
-  
-  
-  
   
 }
